created new Event parent class
