+++ src/main/java/UserElements/Command.java
@@ -1,9 +1,7 @@
 package UserElements;
 
-import Events.EventTypes.Deadline;
 import Events.EventTypes.Event;
 import Events.EventTypes.Task;
-import Events.EventTypes.ToDo;
 import Events.Formatting.DateObj;
 import Events.Storage.Storage;
 import Events.Storage.TaskList;
